requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
colorama==0.4.6
future==0.18.3
httpcore==1.0.4
httplib2==0.14.0
httpx==0.27.0
jsonpatch==1.33
jsonpointer==2.4
keyring==18.0.1
matplotlib==3.1.2
msgpack==1.0.7
networkx==3.1
nltk==3.8.1
numpy==1.21.
oauthlib==3.1.0
pandas==2.0.3
pillow==10.2.0
pipx==1.4.3
psutil==5.5.1
py-trees==0.7.6
pydantic==2.6.3
pydantic-core==2.16.3
pydot==1.4.1
pygame==2.5.2
pyserial==3.4
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.1
redlock==1.2.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
shapely==2.0.3
simplejson==3.16.0
urllib3==2.2.0
zmq==0.0.0
